fix: 팀스페이스 추가 팝업창 띄었을 때, 잘못된 GET api 요청하는 버그 수정
