+++ src/hooks/queries/Team/useTeamLodging.jsx
@@ -1,27 +1,12 @@
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 import teamAPI from '../../../apis/teamAPI';
 
 const useTeamLodging = (userToken, teamToken = {}) => {
-  const queryClient = useQueryClient();
-
   const teamLodgingQuery = useQuery(['team', teamToken, 'lodging'], () =>
     teamAPI.lodgingList(userToken, teamToken),
   );
 
-  const { mutate: teamLodgingMutation } = useMutation(
-    (paylod) => teamAPI.lodgingCreate(paylod, userToken),
-    {
-      onSuccess: (data) => {
-        console.log(data);
-        queryClient.invalidateQueries(['team', teamToken, 'lodging']);
-      },
-      onError: (error) => {
-        console.log(error);
-      },
-    },
-  );
-
-  return { teamLodgingQuery, teamLodgingMutation };
+  return { teamLodgingQuery };
 };
 
 export default useTeamLodging;
